refactor(odyssey-react-mui): create new PickerVirtualizationListboxComponent
